requirements.txt
# K1NGB0B Recon Suite v3.0 - Dependencies
# Cross-platform compatible with macOS and Linux

# Core HTTP and async libraries
aiohttp>=3.8.0
requests>=2.28.0

# DNS resolution
dnspython>=2.3.0

# System utilities
psutil>=5.9.0

# Optional: YAML config support
pyyaml>=6.0
